Move test search filter to own file
