+++ context/app/static/js/components/test-search/filters/Filter/enums.js
@@ -1,0 +1,3 @@
+export const FILTER_TYPES = {
+  multiList: 'multiList',
+};
