+++ context/app/static/js/components/test-search/filters/Filter/index.js
@@ -1,0 +1,3 @@
+import Filter from './Filter';
+
+export default Filter;
